elevator
This is the  advanced development of the elevator. In this code elevator
picks up multiple people. Minimum number of floors for this code are
fixed to 2.
1. First, person thread gets initialised it starts adding people in a
global linked list at certain "arrival time"
2. Each elevator thread then works in following way::
a. Each elevator is kept on first, and last floor alternately for better
functioning
b. Elevator then checks whether there are people in its own list
i) If so and if there are no people in the global list, it starts its
work without applying locks and without accessing the global list
ii)If there is no person it threads own list then it waits for global
list to have atleast one person in it
c. The destination and direction of the linked list is set, and any
other person who wants to go in the same direction and leave the
elevator before or on the same floor as that of the elevators
destination.
d. Then the own list of the elevator is checked to print the person id
who will enter in the elevator, and set the flag for each person who has
entered the elevator
e. The the function to drop the people to certain floor is called then
Depending on the direction of each person and the flag telling that the
person has entered the elevator the person node is removedfrom the
elevator.
3.When the simulation time gets over both elevator and person thread
stops working, and memory will be freed.